requirements.txt
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
cryptography==36.0.1
google-api-core==2.3.2
google-auth==2.3.3
googleapis-common-protos==1.54.0
grpc-requests==0.1.0
grpcio==1.43.0
grpcio-reflection==1.43.0
grpcio-tools==1.43.0
idna==3.3
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
requests==2.26.0
rsa==4.8
six==1.16.0
urllib3==1.26.7
